backend/requirements.txt
flask
flask-cors
pymongo
twilio
python-dotenv
gunicorn